fix low high equity issue

+++ TradeSystem.Data/Models/_Strategies/RiskManagementSetting.cs
@@ -17,8 +17,8 @@ namespace TradeSystem.Data.Models
         public double MaxSwaps { get; set; }
 
 		public int MaxTicketDuration { get; set; }
-		public int OptimumEquity { get; set; }
-		public int AddEq { get; set; }
-		public int WdrawEq { get; set; }
+		public double OptimumEquity { get; set; }
+		public double AddEq { get; set; }
+		public double WdrawEq { get; set; }
 	}
 }
